PulseObject: Add channel_map attribute with raw position integers
This commit adds an attribute `channel_map` to all PulseObject objects
that are build from a C struct with a channel_map field, such as
PulseSinkInfo or PulseSourceInfo. In contrast to `channel_list`, which
is already included with these objects, channel_map is a list of the raw
integer position codes from libpulse's `pa_channel_position` enum instead
of a string representation of the channel positions.